Complete AUTOMINE SaaS platform implementation
- Database: MySQL schema with 15+ tables, relationships, and seeding
- Backend: Spring Boot API with JWT auth, clean architecture, and enterprise modules
- Frontend: Angular 18 web app with Material UI and industrial theming
- Mobile: Flutter app skeleton for employee access
- DevOps: Docker setup with docker-compose for local development
- CI/CD: GitHub Actions workflow for automated builds
- Documentation: Complete project structure and setup guides

+++ automine-platform/backend/src/main/java/com/automine/platform/controller/EmployeeController.java
@@ -32,7 +32,7 @@ public class EmployeeController {
 
     @PatchMapping("/{id}/retire")
     @PreAuthorize("hasAnyRole('ADMIN','RRHH')")
-    public ResponseEntity<Void> retire(@PathVariable Long id) {
+    public ResponseEntity<Void> retire(@PathVariable Integer id) {
         employeeService.retire(id);
         return ResponseEntity.noContent().build();
     }
