+++ automine-platform/backend/src/main/java/com/automine/platform/service/UserService.java
@@ -8,5 +8,5 @@ import java.util.List;
 public interface UserService {
     UserResponse create(CreateUserRequest request);
     List<UserResponse> listActive();
-    void deactivate(Long id);
+    void deactivate(Integer id);
 }
